move visitor based classes to visitor package
