+++ src/main/java/edu/kis/powp/jobs2d/command/visitor/CommandVisitorTransformation.java
@@ -1,4 +1,4 @@
-package edu.kis.powp.jobs2d.command.transformation;
+package edu.kis.powp.jobs2d.command.visitor;
 
 import edu.kis.powp.jobs2d.command.*;
 import edu.kis.powp.jobs2d.command.visitor.CommandVisitorInterface;
